backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
pillow
python-multipart